add proper cli and fix ci
