+++ .github/workflows/ci.yml
@@ -18,14 +18,12 @@ jobs:
       - name: set up python
         uses: actions/setup-python@v4
         with:
-          python-version: 3.10
+          python-version: '3.10'
 
       - uses: actions/cache@v3
         id: cache
         with:
-          path: |
-            ${{ env.pythonLocation }}
-            .mypy_cache
+          path: ${{ env.pythonLocation }}
           key: >
             lint
             ${{ runner.os }}
