+++ dnserver/__main__.py
@@ -1,0 +1,7 @@
+"""
+This allows usage via `python -m dnserver`
+"""
+from .cli import cli
+
+if __name__ == '__main__':
+    cli()
