+++ pyproject.toml
@@ -7,7 +7,7 @@ path = 'dnserver/version.py'
 
 [project]
 name = 'dnserver'
-description = 'Job queues in python with asyncio and redis'
+description = 'Simple DNS server written in python for use in development and testing.'
 authors = [{name = 'Samuel Colvin', email = 's@muelcolvin.com'}]
 license = {file = 'LICENSE'}
 readme = 'README.md'
@@ -90,25 +90,3 @@ include_trailing_comma = true
 force_grid_wrap = 0
 combine_as_imports = true
 color_output = true
-
-[tool.mypy]
-show_error_codes = true
-follow_imports = 'silent'
-strict_optional = true
-warn_redundant_casts = true
-warn_unused_ignores = true
-disallow_any_generics = true
-check_untyped_defs = true
-no_implicit_reexport = true
-warn_unused_configs = true
-disallow_subclassing_any = true
-disallow_incomplete_defs = true
-disallow_untyped_decorators = true
-disallow_untyped_calls = true
-
-# for strict mypy: (this is the tricky one :-))
-disallow_untyped_defs = true
-
-# remaining arguments from `mypy --strict` which cause errors
-#no_implicit_optional = true
-#warn_return_any = true
