+++ requirements/linting.in
@@ -3,6 +3,5 @@ flake8
 flake8-pyproject
 flake8-quotes
 isort[colors]
-mypy
 pyupgrade
 pre-commit
